+++ sporttracker/static/js/mapTileHuntingHeatmap.js
@@ -14,14 +14,14 @@ function initMap()
     let map = initMapBase();
 
     L.tileLayer(tileRenderUrl + '/{z}/{x}/{y}.png', {
-        minZoom: 9,
+        minZoom: mapMinZoomLevel,
         maxZoom: 16
     }).addTo(map);
 
     map.on('zoomend', function(e)
     {
         const currentZoom = map.getZoom();
-        document.getElementById('warning-zoom').classList.toggle('d-none', currentZoom >= 9);
+        document.getElementById('warning-zoom').classList.toggle('d-none', currentZoom >= mapMinZoomLevel);
     });
 
     map.on('click', function(e)